code-generation/requirements.txt
yapf>=0.40.2
pyyaml>=6.0.2
